+++ src/constants.php
@@ -8,6 +8,8 @@ namespace Rubix\Server\Http
 
     const UNAUTHORIZED = 401;
 
+    const FORBIDDEN = 403;
+
     const NOT_FOUND = 404;
 
     const METHOD_NOT_ALLOWED = 405;