+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - 0.2.3-beta
     - Non-error responses deferred for better concurrency
     - Rename internal command Response objects to Payload
+    - Trusted Clients middleware now returns Forbidden (403) response
 
 - 0.2.2-beta
     - Added Async Client interface